pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-polymorphic>=3.1.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "setuptools>=76.0.0",
]
